% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{package}
\name{templates}
\alias{templates}
\alias{templates-package}
\title{Rmarkdown templates}
\description{
Rmarkdown templates for SOE and other reports. After installing the package,
click on `create new Rmarkdown document` in Rstudio to have access to the templates.
See \url{http://rmarkdown.rstudio.com/developer_document_templates.html} for
more information.
}
\author{
Scott Large \email{scott.large@noaa.gov}
}
\keyword{package}
